refactor: db 저장을 위한 객체들의 패키지로 이동
